Reverted the reel spinning to use floating point calculations; Improved reeling logicand visuals

+++ Singleton/headers/Singleton_Common.hpp
@@ -18,7 +18,7 @@ namespace single {
     };
 
     struct Rect {
-        int x, y, w, h;
+        float x, y, w, h;
     };
 
     struct SDLDeleter {
